[cache] Adding local cache support
Adding local cache support to git-got. This cache is per-user based which
means if the user has many copies of repositories that point to the same
git-got files the cache can be shared.

The way this is handled by adding a decorator to the load function which
when both the cache is available and enabled will first try to retrieve the
file from the cache, and if not available will handle the regular load. Once
the regular load is done if again both cache available and enabled the
file is stored in the cache by checksum (which by default is sha-256 AFAIK)

clear_local_cache has been added to wipe the content from the cache

-C and --no-cache are added as options to disable the cache